+++ modules/data/spec/dataservices/entity-data.service.spec.ts
@@ -7,11 +7,6 @@ import { Observable } from 'rxjs';
 import { Update } from '@ngrx/entity';
 
 import {
-  createEntityDefinition,
-  EntityDefinition,
-  EntityMetadata,
-  EntityMetadataMap,
-  ENTITY_METADATA_TOKEN,
   DefaultDataService,
   DefaultDataServiceFactory,
   HttpUrlGenerator,