+++ modules/data/spec/reducers/entity-cache-reducer.spec.ts
@@ -28,7 +28,6 @@ import {
   ChangeSet,
   ChangeSetOperation,
   Logger,
-  EntityMetadata,
 } from '../..';
 
 class Hero {