+++ modules/data/src/actions/entity-action-operators.ts
@@ -1,7 +1,4 @@
-import { Action } from '@ngrx/store';
-import { Actions } from '@ngrx/effects';
-
-import { Observable, OperatorFunction } from 'rxjs';
+import { OperatorFunction } from 'rxjs';
 import { filter } from 'rxjs/operators';
 
 import { EntityAction } from './entity-action';