+++ modules/data/src/dataservices/persistence-result-handler.service.ts
@@ -1,15 +1,13 @@
 import { Injectable } from '@angular/core';
 import { Action } from '@ngrx/store';
 
-import { Observable, of } from 'rxjs';
-
 import {
   DataServiceError,
   EntityActionDataServiceError,
 } from './data-service-error';
 import { EntityAction } from '../actions/entity-action';
 import { EntityActionFactory } from '../actions/entity-action-factory';
-import { EntityOp, makeErrorOp, makeSuccessOp } from '../actions/entity-op';
+import { makeErrorOp, makeSuccessOp } from '../actions/entity-op';
 import { Logger } from '../utils/interfaces';
 
 /**