+++ modules/data/src/reducers/constants.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
-import { Action, ActionReducer, MetaReducer } from '@ngrx/store';
+import { MetaReducer } from '@ngrx/store';
 import { EntityCache } from './entity-cache';
 
 export const ENTITY_CACHE_NAME = 'entityCache';