+++ modules/store/src/selector.ts
@@ -1,4 +1,5 @@
 import { Selector, SelectorWithProps } from './models';
+import { isDevMode } from '@angular/core';
 
 export type AnyFn = (...args: any[]) => any;
 
@@ -94,9 +95,9 @@ export function defaultMemoize(
       return lastResult;
     }
 
+    const newResult = projectionFn.apply(null, arguments as any);
     lastArguments = arguments;
 
-    const newResult = projectionFn.apply(null, arguments as any);
     if (isResultEqual(lastResult, newResult)) {
       return lastResult;
     }
@@ -274,6 +275,7 @@ export function createSelector<State, S1, S2, S3, S4, S5, S6, Result>(
     Selector<State, S1>,
     Selector<State, S2>,
     Selector<State, S3>,
+
     Selector<State, S4>,
     Selector<State, S5>,
     Selector<State, S6>
@@ -601,8 +603,19 @@ export function createFeatureSelector<T, V>(
 export function createFeatureSelector(
   featureName: any
 ): MemoizedSelector<any, any> {
-  return createSelector(
-    (state: any) => state[featureName],
-    (featureState: any) => featureState
-  );
+  return createSelector((state: any) => {
+    const featureState = state[featureName];
+    if (isDevMode() && featureState === undefined) {
+      console.warn(
+        `The feature name \"${featureName}\" does ` +
+          'not exist in the state, therefore createFeatureSelector ' +
+          'cannot access it.  Be sure it is imported in a loaded module ' +
+          `using StoreModule.forRoot('${featureName}', ...) or ` +
+          `StoreModule.forFeature('${featureName}', ...).  If the default ` +
+          'state is intended to be undefined, as is the case with router ' +
+          'state, this development-only warning message can be ignored.'
+      );
+    }
+    return featureState;
+  }, (featureState: any) => featureState);
 }